Kanji button sets search param now, plus refactors
